+++ CHANGELOG.md
@@ -6,4 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## Added
+
+- **! Added `External` progress bar**
+- Added microseconds and nanoseconds
+
 ## [0.1.0] - 2022-10-28